requirements.txt
django
pymongo
python-decouple
openai